--- src/app/_models/asset.ts
@@ -1,8 +1,0 @@
-export interface Asset {
-   id: number;
-   assetName: string;
-   assetType: string;
-   age: number;
-   purchasePrice: number;
-   // proofOfOwnership: string;
-}
