+++ src/app/pages/user-pages/login-page/login-page.component.ts
@@ -15,13 +15,6 @@ import { ApiService } from 'src/app/_services/api-service/api.service';
 import { Subscription } from 'rxjs';
 import { AuthService } from 'src/app/_services/authentication-service/auth.service';
 import { ValidationService } from 'src/app/_services/validation-service/validation.service';
-import { FarmService } from 'src/app/_services/farm-service/farm.service';
-import { Plot } from 'src/app/_models/plot';
-import { PlotService } from 'src/app/_services/plot-service/plot.service';
-import { CropService } from 'src/app/_services/crop-service/crop.service';
-import { Assets } from 'src/app/_models/Assets';
-import { AssetService } from 'src/app/_services/asset-service/asset.service';
-import { JwtService } from 'src/app/_services/JWT-service/jwt.service';
 
 @Component({
     selector: 'app-login-page',
@@ -33,7 +26,6 @@ export class LoginPageComponent implements OnInit, OnDestroy {
     message = ''; // Message to display error or validation messages
     // used to store subscriptions to services
     private subscription = new Subscription();
-    apiResponse: any;
 
     constructor(
         private fb: FormBuilder,
@@ -41,11 +33,6 @@ export class LoginPageComponent implements OnInit, OnDestroy {
         private router: Router,
         // private userService: UserService, // Service to handle user-related functionalities
         private _apiService: ApiService, // Service to make API requests
-        private _farmService: FarmService,
-        private _plotService: PlotService,
-        private _cropService: CropService,
-        private _assetService: AssetService,
-        private _jwtService: JwtService,
         private _authService: AuthService
     ) {}
 
@@ -85,54 +72,4 @@ export class LoginPageComponent implements OnInit, OnDestroy {
             );
         }
     }
-    // logs a user into the application
-    loginUser(loginEmail: string, loginPassword: string) {
-        // set the body that will be passed to the api connection
-        const loginBody = {
-            email: loginEmail,
-            password: loginPassword,
-        };
-        this._apiService.loginUser(loginBody).subscribe((result: any) => {
-            // assigns the result to the structure of the api response object
-            this.apiResponse = result;
-            // sets the token to the one received from the api
-            this._jwtService.setToken(this.apiResponse.data);
-            // sets the user login state to true
-            this.setUserState();
-            if (this._farmService.getFarmInfo().length > 0) {
-                // routes to dashboard if the login was successful
-                this.router.navigate(['/dashboard']);
-            } else {
-                // TODO: if the farm, crop, plot, asset info is blank it will route to tell me about your farm
-                this.router.navigate(['/about-farm']);
-            }
-            if (this._plotService.getPlotInfo().length > 0) {
-                // routes to dashboard if the login was successful
-                this.router.navigate(['/dashboard']);
-            } else {
-                // TODO: if the farm, crop, plot, asset info is blank it will route to tell me about your farm
-                this.router.navigate(['/about-farm']);
-            }
-            this._cropService.getCropInfo().subscribe(crops => {
-                if (crops.length > 0) {
-                    this.router.navigate(['/dashboard']);
-                } else {
-                    this.router.navigate(['/about-farm']);
-                }
-            });
-
-            this._assetService.getAssetInfo().subscribe(assets => {
-                if (assets.length > 0) {
-                    this.router.navigate(['/dashboard']);
-                } else {
-                    this.router.navigate(['/about-farm']);
-                }
-            });
-        });
-    }
-    setUserState() {
-        throw new Error('Method not implemented.');
-    }
-
-  
 }
